// backup/graphql-types.ts
export type Maybe<T> = T | null;
// ...
export type Wordpress_Category = Node & Wordpress_Node & Wordpress_TermNode & Wordpress_UniformResourceIdentifiable & Wordpress_DatabaseIdentifier & Wordpress_HierarchicalTermNode & Wordpress_MenuItemLinkable & {
  /** The ancestors of the node. Default ordered as lowest (closest to the child) to highest (closest to the root). */
  ancestors?: Maybe<Wordpress_CategoryToAncestorsCategoryConnection>;
  /** Connection between the category type and the category type */
  wpChildren?: Maybe<Wordpress_CategoryToCategoryConnection>;
  /** Connection between the category type and the ContentNode type */
  contentNodes?: Maybe<Wordpress_CategoryToContentNodeConnection>;
  /** The number of objects connected to the object */
  count?: Maybe<Scalars['Int']>;
  /** The unique resource identifier path */
  databaseId: Scalars['Int'];
  /** The description of the object */
  description?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean'];
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean'];
  /** The link to the term */
  link?: Maybe<Scalars['String']>;
  /** The human friendly name of the object. */
  name?: Maybe<Scalars['String']>;
  /** Connection between the category type and the page type */
  pages?: Maybe<Wordpress_CategoryToPageConnection>;
  /** Connection between the category type and the category type */
  wpParent?: Maybe<Wordpress_CategoryToParentCategoryConnectionEdge>;
  /** Database id of the parent node */
  parentDatabaseId?: Maybe<Scalars['Int']>;
  /** The globally unique identifier of the parent node. */
  parentId?: Maybe<Scalars['ID']>;
  /** Connection between the category type and the post type */
  posts?: Maybe<Wordpress_CategoryToPostConnection>;
  /** An alphanumeric identifier for the object unique to its type. */
  slug?: Maybe<Scalars['String']>;
  /** Connection between the category type and the Taxonomy type */
  taxonomy?: Maybe<Wordpress_CategoryToTaxonomyConnectionEdge>;
  /** The name of the taxonomy that the object is associated with */
  taxonomyName?: Maybe<Scalars['String']>;
  /** The ID of the term group that this term object belongs to */
  termGroupId?: Maybe<Scalars['Int']>;
  /** The taxonomy ID that the object is associated with */
  termTaxonomyId?: Maybe<Scalars['Int']>;
  /** The unique resource identifier path */
  uri?: Maybe<Scalars['String']>;
  nodeType?: Maybe<Scalars['String']>;
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type Wordpress_Post_Postacf = {
  division?: Maybe<Scalars['String']>;
  fieldGroupName?: Maybe<Scalars['String']>;
  flashPost?: Maybe<Wordpress_Post_Postacf_FlashPost>;
  matchReport?: Maybe<Wordpress_Post_Postacf_MatchReport>;
  normalPost?: Maybe<Wordpress_Post_Postacf_NormalPost>;
  postCategory?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Wordpress_Post_Postacf_FlashPost = {
  end?: Maybe<Scalars['String']>;
  fieldGroupName?: Maybe<Scalars['String']>;
  start?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Wordpress_Post_Postacf_MatchReport = {
  assists?: Maybe<Scalars['String']>;
  fieldGroupName?: Maybe<Scalars['String']>;
  gameDay?: Maybe<Scalars['String']>;
  penalties?: Maybe<Scalars['String']>;
  /** Die Drittelergebniss als Text eingeben. Beispiel (1:0, 0:0, 2:2, 1:0) n.P. */
  periodResults?: Maybe<Scalars['String']>;
  scoreGuest?: Maybe<Scalars['Float']>;
  scoreHome?: Maybe<Scalars['Float']>;
  scorer?: Maybe<Scalars['String']>;
  /** Wurde das Spiel in der Nachspielzeit oder nach Penalty schießen beendet? Wurde es Abgebrochen? */
  specialCase?: Maybe<Scalars['String']>;
  spectators?: Maybe<Scalars['String']>;
  teamGuest?: Maybe<Scalars['String']>;
  teamGuestSlug?: Maybe<Scalars['String']>;
  teamHome?: Maybe<Scalars['String']>;
  teamHomeSlug?: Maybe<Scalars['String']>;
};
// ...
export type Wordpress_Page_Pageacf = {
  boardMembers?: Maybe<Array<Maybe<Wordpress_Page_Pageacf_BoardMembers>>>;
  /** Bitte auswählen für welchen Bereich dieser Eintrag gelten soll. Default ist &#039;verein&#039; */
  division?: Maybe<Scalars['String']>;
  fieldGroupName?: Maybe<Scalars['String']>;
  gamepitchWidget?: Maybe<Scalars['String']>;
  lineup?: Maybe<Wordpress_Page_Pageacf_Lineup>;
  seo?: Maybe<Wordpress_Page_Pageacf_Seo>;
  shop?: Maybe<Wordpress_Page_Pageacf_Shop>;
  sponsorsContent?: Maybe<Array<Maybe<Wordpress_Page_Pageacf_SponsorsContent>>>;
  standardContent?: Maybe<Wordpress_Page_Pageacf_StandardContent>;
  teamContent?: Maybe<Wordpress_Page_Pageacf_TeamContent>;
  /** Bitte das Template für die Seite auswählen */
  template?: Maybe<Scalars['String']>;
  ticketsContent?: Maybe<Wordpress_Page_Pageacf_TicketsContent>;
};
// ...
export type PageQuery = { wordpressPage?: Maybe<PageFieldsFragment>, allWordpressPost: { nodes: Array<BlogPostPreviewFieldsFragment> } };
// ...
export type PageFieldsFragment = (
  Pick<Wordpress_Page, 'slug' | 'title' | 'id' | 'content' | 'uri'>
  & { pageACF?: Maybe<(
    Pick<Wordpress_Page_Pageacf, 'template' | 'division' | 'gamepitchWidget'>
    & { standardContent?: Maybe<StandardContentFieldsFragment>, teamContent?: Maybe<TeamContentFieldsFragment>, sponsorsContent?: Maybe<Array<Maybe<SponsorsContentFieldsFragment>>>, ticketsContent?: Maybe<TicketsContentFieldsFragment>, lineup?: Maybe<{ teamPicture?: Maybe<(
        Pick<Wordpress_MediaItem, 'altText' | 'title'>
        & { localFile?: Maybe<{ childImageSharp?: Maybe<{ fluid?: Maybe<FluidWithWebpFieldsFragment> }> }> }
      )>, team?: Maybe<Array<Maybe<PlayerFieldsFragment>>> }>, boardMembers?: Maybe<Array<Maybe<BoardMemberFieldsFragment>>> }
    & PageAcfSeoFieldsFragment
  )> }
  & HomeAcfFieldsFragment
);

export type BlogPostPreviewFieldsFragment = (
  Pick<Wordpress_Post, 'title' | 'excerpt' | 'date' | 'id' | 'uri'>
  & { postACF?: Maybe<(
    Pick<Wordpress_Post_Postacf, 'division' | 'postCategory'>
    & { flashPost?: Maybe<Pick<Wordpress_Post_Postacf_FlashPost, 'start' | 'end'>>, matchReport?: Maybe<Pick<Wordpress_Post_Postacf_MatchReport, 'assists' | 'scoreGuest' | 'scoreHome'>> }
  )>, categories?: Maybe<{ nodes?: Maybe<Array<Maybe<Pick<Wordpress_Category, 'name'>>>> }>, featuredImage?: Maybe<{ node?: Maybe<(
      Pick<Wordpress_MediaItem, 'altText'>
      & { localFile?: Maybe<{ childImageSharp?: Maybe<{ fluid?: Maybe<FluidWithWebpFieldsFragment> }> }> }
    )> }> }
);

export type StandardContentFieldsFragment = (
  Pick<Wordpress_Page_Pageacf_StandardContent, 'pageContentTitle'>
  & { contact?: Maybe<Pick<Wordpress_Page_Pageacf_StandardContent_Contact, 'email' | 'name' | 'tel' | 'website'>>, additionalInfo?: Maybe<Pick<Wordpress_Page_Pageacf_StandardContent_AdditionalInfo, 'title' | 'content'>>, downloads?: Maybe<Array<Maybe<{ file?: Maybe<Pick<Wordpress_MediaItem, 'title' | 'mediaItemUrl'>> }>>> }
  & FeaturePicFieldsFragment
);

export type TicketsContentFieldsFragment = (
  Pick<Wordpress_Page_Pageacf_TicketsContent, 'season' | 'suffix'>
  & { standing?: Maybe<Array<Maybe<Pick<Wordpress_Page_Pageacf_TicketsContent_Standing, 'kind' | 'singleCost' | 'seasonpassCost'>>>>, seating?: Maybe<Array<Maybe<Pick<Wordpress_Page_Pageacf_TicketsContent_Seating, 'kind' | 'singleCost' | 'seasonpassCost'>>>>, seasonpass?: Maybe<(
    Pick<Wordpress_Page_Pageacf_TicketsContent_Seasonpass, 'description'>
    & { applications?: Maybe<Array<Maybe<(
      Pick<Wordpress_Page_Pageacf_TicketsContent_Seasonpass_Applications, 'title'>
      & { application?: Maybe<Pick<Wordpress_MediaItem, 'title' | 'mediaItemUrl'>> }
    )>>> }
  )> }
);

export type PlayerFieldsFragment = (
  Pick<Wordpress_Page_Pageacf_Lineup_Team, 'name' | 'nationality' | 'number' | 'position' | 'rodiLink' | 'stickSide'>
  & { portrait?: Maybe<(
    Pick<Wordpress_MediaItem, 'altText'>
    & { localFile?: Maybe<{ childImageSharp?: Maybe<{ fluid?: Maybe<FluidWithWebpFieldsFragment> }> }> }
  )> }
);

export type BoardMemberFieldsFragment = (
  Pick<Wordpress_Page_Pageacf_BoardMembers, 'name' | 'level' | 'position' | 'email' | 'telephone'>
  & { portrait?: Maybe<(
    Pick<Wordpress_MediaItem, 'altText'>
    & { localFile?: Maybe<{ childImageSharp?: Maybe<{ fluid?: Maybe<FluidWithWebpFieldsFragment> }> }> }
  )> }
);

export type TeamContentFieldsFragment = (
  Pick<Wordpress_Page_Pageacf_TeamContent, 'showTable'>
  & { teamPicture?: Maybe<(
    Pick<Wordpress_MediaItem, 'altText' | 'title'>
    & { localFile?: Maybe<{ childImageSharp?: Maybe<{ fluid?: Maybe<FluidWithWebpFieldsFragment> }> }> }
  )>, ranking?: Maybe<Array<Maybe<RankingFieldsFragment>>>, team?: Maybe<{ teamMembers?: Maybe<Array<Maybe<TeamMemberFieldsFragment>>>, contact?: Maybe<Pick<Wordpress_Page_Pageacf_TeamContent_Team_Contact, 'coach' | 'assistant' | 'email' | 'manager' | 'telephone'>> }> }
);

export type TeamMemberFieldsFragment = Pick<Wordpress_Page_Pageacf_TeamContent_Team_TeamMembers, 'pos' | 'number' | 'name'>;

export type RankingFieldsFragment = Pick<Wordpress_Page_Pageacf_TeamContent_Ranking, 'team' | 'points' | 'goalsAgainst' | 'goals' | 'gamesAmount' | 'draws' | 'defeats' | 'wins'>;

export type SponsorsContentFieldsFragment = (
  Pick<Wordpress_Page_Pageacf_SponsorsContent, 'name' | 'level' | 'url'>
  & { logo?: Maybe<(
    Pick<Wordpress_MediaItem, 'altText' | 'title'>
    & { localFile?: Maybe<{ childImageSharp?: Maybe<{ fluid?: Maybe<FluidWithWebpFieldsFragment> }> }> }
  )> }
);

export type FluidWithWebpFieldsFragment = Pick<ImageSharpFluid, 'src' | 'sizes' | 'srcWebp' | 'srcSet' | 'aspectRatio'>;
// ...
export type PageAcfSeoFieldsFragment = { seo?: Maybe<Pick<Wordpress_Page_Pageacf_Seo, 'metaDescription' | 'noIndex' | 'ogDescription' | 'twitterDescription'>> };

export type HomeAcfFieldsFragment = { homeACF?: Maybe<Pick<Wordpress_Page_Homeacf, 'adresse'>> };